+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ---
 
+# 1.0.5 — 2026-03-17
+
+Date/value columns & 'event_type'/'CDW_type' columns in results, outcome argument values ('custom').
+
+### Changed
+- Functions MSprog(), value_milestone(), separate_ri_ra(), compute_delta(), is_event()
+
+---
+
 # 1.0.4 — 2026-02-05
 
 Included date_format argument.